% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developr-package.R
\name{access_rda}
\alias{access_rda}
\title{Copy Microsoft Access tables to an R package data/ folder.}
\usage{
access_rda(access_path, data_path, tables = NULL, tables_rda = NULL)
}
\arguments{
\item{access_path}{Microsoft Access database path.}

\item{data_path}{R packaga data path.}

\item{tables}{Optional selection of tables to copy.}

\item{tables_rda}{Optional tables names to save in data/.}
}
\description{
Copy Microsoft Access tables to an R package data/ folder.
}
