% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{get_root}
\alias{get_root}
\title{Get root directory}
\usage{
get_root(dir_test = "C:/OVE/")
}
\arguments{
\item{dir_test}{Directory to test exisence in order to specify the root directory.}
}
\value{
The root directory
}
\description{
Get root directory
}
