% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developr.R
\name{package_build}
\alias{package_build}
\title{Build a package.}
\usage{
package_build(package_path = ".", documentation = TRUE, ...)
}
\arguments{
\item{package_path}{Package path.}

\item{documentation}{If \code{TRUE}, documentation is also generated.}

\item{\dots}{Additional arguments to \code{devtools::install_local()}.}
}
\description{
Build a package.
}
