% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developr.R
\name{package_deps}
\alias{package_deps}
\title{Get package dependancies.}
\usage{
package_deps(package_path = getwd(), include_rmd = TRUE)
}
\arguments{
\item{package_path}{Package path.}

\item{include_rmd}{If \code{TRUE} then find dependances into Rmd files as well.}
}
\value{
A dependencies vector.
}
\description{
Find packages dependances declared into R scripts and Rmd files.
}
