% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{replace_non_ascii_with_unicode}
\alias{replace_non_ascii_with_unicode}
\title{Replace non ASCII character to unicode.}
\usage{
replace_non_ascii_with_unicode(r_file, r_file_output = NULL)
}
\arguments{
\item{r_file}{R file.}

\item{r_file_output}{Output R file.}
}
\description{
Replace non ASCII character to unicode.
}
